refactor: 로고 img 컴포넌트 리팩터링
