+++ next-app/src/components/common/FeedItem/FeedItem.utils.ts
@@ -6,6 +6,14 @@ export const copyToClipboard = async (linkUrl: string) => {
     Toast.show({ content: '클립보드에 복사되었습니다.' })
   } catch (error) {
     Toast.show({ type: 'error', content: '복사에 실패하였습니다.' })
-    console.log(error)
+  }
+}
+
+export const getDiameterByType = (type: string) => {
+  if (type.includes('card')) {
+    return 20
+  }
+  if (type.includes('subscription')) {
+    return 48
   }
 }
